requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
duckdb==1.2.2
groq==0.4.2
fastapi==0.110.0
uvicorn==0.27.1
prefect==2.16.4
requests==2.31.0
python-dotenv==1.0.1
pandas
pyarrow
cryptography
feedparser==6.0.11
trafilatura
vaderSentiment==3.3.2
curl_cffi
beautifulsoup4
setuptools
sentence-transformers
chromadb
google-generativeai
python-dotenv
